+++ hoot-core/src/main/cpp/hoot/core/util/PositiveIdGenerator.cpp
@@ -51,9 +51,9 @@ IdGeneratorPtr PositiveIdGenerator::clone() const
 
 void PositiveIdGenerator::reset()
 {
-  _wayId = ConfigOptions().getIdGeneratorWayStart();
-  _nodeId = ConfigOptions().getIdGeneratorNodeStart();
-  _relationId = ConfigOptions().getIdGeneratorRelationStart();
+  _wayId = (long)ConfigOptions().getIdGeneratorWayStart();
+  _nodeId = (long)ConfigOptions().getIdGeneratorNodeStart();
+  _relationId = (long)ConfigOptions().getIdGeneratorRelationStart();
 }
 
 }
